JSight to OpenAPI converter, const rule for primitives
