+++ openapi/internal/jsoac/jsoac_test.go
@@ -81,76 +81,6 @@ func TestNewOpenAPI(t *testing.T) {
 				}
 				}`,
 			},
-
-			{
-				`
-					"test@test.ru" //{ type: "email" }
-				`,
-				`
-					{
-						"type": "string", 
-						"example": "test@test.ru",
-						"format":"email"
-					}
-				`,
-			},
-			{
-				`
-					"https://www.com" //{ type: "uri" }
-				`,
-				`
-					{
-						"type": "string", 
-						"example": "https://www.com",
-						"format":"uri"
-					}
-				`,
-			},
-			{
-				`
-					"53496d7f-1374-4368-a829-74ccd47aec1c" //{ type: "uuid" }
-				`,
-				`
-					{
-						"type": "string", 
-						"example": "53496d7f-1374-4368-a829-74ccd47aec1c",
-						"format":"uuid"
-					}
-				`,
-			},
-			{
-				`
-					"2024-02-14" //{ type: "date" }
-				`,
-				`
-					{
-						"type": "string", 
-						"example": "2024-02-14",
-						"format":"date"
-					}
-				`,
-			},
-			{
-				`
-					"2024-02-14T09:14:28+03:00" //{ type: "datetime" }
-				`,
-				`
-					{
-						"type": "string", 
-						"example": "2024-02-14T09:14:28+03:00",
-						"format":"date-time"
-					}
-				`,
-			},
-			//TODO: support for different integer formats, such as int32 and int64
-			/*{
-				`123 // {type: "integer"}`,
-				`{"type": "integer", "example": 123, "format": "int32"}`,
-			},*/
-			{
-				`123 // {type: "integer"}`,
-				`{"type": "integer", "example": 123, "format": "int32"}`,
-			},
 		}
 
 		for _, tt := range tests {
